requirements.txt
alembic==1.12.1
asgiref==3.8.1
blinker==1.6.3
click==8.1.7
Django==5.1.4
djangorestframework==3.15.2
Flask==3.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
itsdangerous==2.1.2
Mako==1.2.4
psutil==6.1.0
PyMySQL==1.1.0
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
Werkzeug==3.0.1
